Adds functionality for creation/movement/deletion of directories.
When moving a directory with files in it, it moves the dir first (which also moves the files), but then watchdog attempts to move the already moved files. Not a huge problem, since flask will just give an IOError and say the files don't exist and continue operation, but something to look into later.